clean: move console DSL
