+++ src/main/scala/com/nulabinc/backlog/c2b/dsl/ConsoleADT.scala
@@ -1,0 +1,6 @@
+package com.nulabinc.backlog.c2b.dsl
+
+sealed trait ConsoleADT[A]
+case class Print(str: String) extends ConsoleADT[Unit]
+case class PrintBold(str: String) extends ConsoleADT[Unit]
+case class Read(printMessage: String) extends ConsoleADT[String]
