+++ src/main/scala/com/nulabinc/backlog/c2b/services/CybozuStore.scala
@@ -5,10 +5,9 @@ import java.io.File
 import com.nulabinc.backlog.c2b.core.Logger
 import com.nulabinc.backlog.c2b.datas._
 import com.nulabinc.backlog.c2b.datas.Types.AnyId
-import com.nulabinc.backlog.c2b.dsl.AppDSL
+import com.nulabinc.backlog.c2b.dsl.{AppDSL, ConsoleDSL}
 import com.nulabinc.backlog.c2b.dsl.AppDSL.AppProgram
 import com.nulabinc.backlog.c2b.exceptions.CybozuLiveImporterException
-import com.nulabinc.backlog.c2b.interpreters.ConsoleDSL
 import com.nulabinc.backlog.c2b.parsers.TextFileParser
 import com.nulabinc.backlog.c2b.persistence.dsl._
 import com.nulabinc.backlog.c2b.persistence.dsl.StoreDSL.StoreProgram
