--- src/test/scala/com/nulabinc/backlog/c2b/parsers/UserParserSpec.scala
@@ -1,14 +1,0 @@
-package com.nulabinc.backlog.c2b.parsers
-
-import com.nulabinc.backlog.c2b.datas.CybozuCSVUser
-import org.scalatest.{FlatSpec, Matchers}
-
-//class UserParserSpec extends FlatSpec with Matchers {
-//
-//  "UserParser" should "parse string" in {
-//    val str = "Shoma Nishi"
-//    val actual = UserParser.toUser(str)
-//    val expect = CybozuCSVUser(firstName = "Shoma", lastName = "Nishi")
-//    actual shouldEqual Right(expect)
-//  }
-//}
